* localaddr.c (krb5_os_localaddr_profile): Cast argument to 	isspace() to int.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13796 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/os/localaddr.c
@@ -850,7 +850,7 @@ krb5_os_localaddr_profile (krb5_context context, struct localaddr_data *datap)
 #endif
 
 	for (cp = *iter, next = 0; *cp; cp = next) {
-	    while (isspace (*cp) || *cp == ',')
+	    while (isspace ((int) *cp) || *cp == ',')
 		cp++;
 	    if (*cp == 0)
 		break;
@@ -859,7 +859,7 @@ krb5_os_localaddr_profile (krb5_context context, struct localaddr_data *datap)
 	    fprintf (stderr, "    addr found in '%s'\n", cp);
 #endif
 	    this = cp;
-	    while (*cp != 0 && !isspace(*cp) && *cp != ',')
+	    while (*cp != 0 && !isspace((int) *cp) && *cp != ',')
 		cp++;
 	    if (*cp != 0) {
 		next = cp + 1;
